chore: add validator and field value types
