+++ core/field_label.ts
@@ -14,18 +14,17 @@ import * as goog from '../closure/goog/goog.js';
 goog.declareModuleId('Blockly.FieldLabel');
 
 import * as dom from './utils/dom.js';
-import {FieldConfig, Field} from './field.js';
+import {Field, FieldConfig} from './field.js';
 import * as fieldRegistry from './field_registry.js';
 import * as parsing from './utils/parsing.js';
 import type {Sentinel} from './utils/sentinel.js';
-
 
 /**
  * Class for a non-editable, non-serializable text field.
  *
  * @alias Blockly.FieldLabel
  */
-export class FieldLabel extends Field {
+export class FieldLabel extends Field<string> {
   /** The html class name to use for this field. */
   private class_: string|null = null;
