+++ core/field_multilineinput.ts
@@ -15,7 +15,7 @@ goog.declareModuleId('Blockly.FieldMultilineInput');
 import * as Css from './css.js';
 import {Field, UnattachedFieldError} from './field.js';
 import * as fieldRegistry from './field_registry.js';
-import {FieldTextInputConfig, FieldTextInput} from './field_textinput.js';
+import {FieldTextInput, FieldTextInputConfig, FieldTextInputValidator} from './field_textinput.js';
 import * as aria from './utils/aria.js';
 import * as dom from './utils/dom.js';
 import {KeyCodes} from './utils/keycodes.js';
@@ -25,6 +25,7 @@ import {Svg} from './utils/svg.js';
 import * as userAgent from './utils/useragent.js';
 import * as WidgetDiv from './widgetdiv.js';
 
+export type FieldMultilineInputValidator = FieldTextInputValidator;
 
 /**
  * Class for an editable text area field.
@@ -65,7 +66,7 @@ export class FieldMultilineInput extends FieldTextInput {
    * for a list of properties this parameter supports.
    */
   constructor(
-      opt_value?: string|Sentinel, opt_validator?: Function,
+      opt_value?: string|Sentinel, opt_validator?: FieldMultilineInputValidator,
       opt_config?: FieldMultilineInputConfig) {
     super(Field.SKIP_SETUP);
